Move handlers dir
